DubaiBot/requirements.txt
python-telegram-bot
SQLAlchemy
psycopg2-binary
python-dotenv
requests 